Add common test module for stdlib tests
