+++ stdlib/test-common/src/test/resources/datafiles/file.xml
@@ -1,0 +1,1 @@
+<name>Ballerina xml file part</name>
