+++ stdlib/test-common/src/test/resources/datafiles/test.tmp
@@ -1,0 +1,1 @@
+Ballerina binary file part
